DTO manualy added using BeansUtils mapper

+++ src/main/java/com/marcode/ebanking/entities/CurrentAccount.java
@@ -14,5 +14,5 @@ import javax.persistence.Entity;
 @AllArgsConstructor
 public class CurrentAccount extends  BankAccount{
 
-    private Long overdraft;
+    private double overdraft;
 }
